gamemanager & shop
move to cpp
